Move server code to separate module
